+++ src/containers/App.js
@@ -5,14 +5,6 @@ import zhLocaleData from 'react-intl/locale-data/zh'
 // import locale data
 import { addLocaleData, IntlProvider } from 'react-intl'
 import { connect } from 'react-redux'
-import Layout from '../components/Layout'
-
-// lodash
-import get from 'lodash/get'
-
-const _ = {
-  get
-}
 
 addLocaleData(enLocaleData)
 addLocaleData(zhLocaleData)
@@ -35,16 +27,9 @@ class App extends PureComponent {
   }
 
   render() {
-    const pathname = _.get(this.props, 'location.pathname')
-
     return (
       <IntlProvider locale={currentLocale} defaultLocale="zh-Hant">
-        <Layout
-          header={this.props.header}
-          pathname={pathname}
-        >
-          {this.props.children}
-        </Layout>
+        {this.props.children}
       </IntlProvider>
     )
   }